Optimisation + New Test
Take into acount the email that haven't created an account. The implementation of the time out is questionable for this edge case.

+++ app/core/users/cruds_users.py
@@ -141,7 +141,7 @@ async def get_user_by_email(
 async def get_recovery_request_within_delay(
     db: AsyncSession,
     email: str,
-    minimumDelayMinutes: float,
+    minimumDelayMinutes: int,
 ) -> models_users.CoreUserRecoverRequest | None:
     result = await db.execute(
         select(models_users.CoreUserRecoverRequest).where(
